Modernize layer to use new primitives
