+++ tactics/ruby.py
@@ -1,0 +1,12 @@
+
+from charmtools.build.tactics import ExactMatch, Tactic
+
+
+class DependenciesTxtTactic(ExactMatch, Tactic):
+    FILENAME = 'dependencies.txt'
+    kind = "dynamic"
+
+    def __call__(self):
+        print('DIE FUCKER')
+        raise Exception('dependencies.txt is deprecated see '
+                        'https://github.com/battlemidget/juju-layer-ruby/pull/8')
