+++ tox.ini
@@ -1,14 +1,14 @@
 [tox]
-envlist = flake, pep8
+envlist = flake8, pep8
 skipsdist = True
 
 [testenv]
 deps = -r{toxinidir}/requirements.txt
 
-[testenv:flake]
-commands = flake8 {posargs} reactive
-deps = flakes
+[testenv:flake8]
+commands = flake8 {posargs} reactive lib/charms
+deps = flake8
 
 [testenv:pep8]
-commands = pep8 {posargs} reactive
+commands = pep8 {posargs} reactive lib/charms
 deps = pep8
